quora-automation-project/requirements.txt
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
urllib3==2.1.0
openai==1.3.9
textblob==0.17.1
nltk==3.8.1
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
APScheduler==3.10.4
python-dotenv==1.0.0
PyJWT==2.8.1
Flask==2.3.0
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
